Python/HA_State/requirements.txt
python-benedict==0.33.2
requests==2.31.0
xmltodict==0.13.0
